% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_gestDM_deterministic_Patch}
\alias{oneDay_gestDM_deterministic_Patch}
\title{Deterministic Gestating Pup Death and Maturation}
\usage{
oneDay_gestDM_deterministic_Patch()
}
\description{
Daily gestating pup survival is calculated according to \deqn{\overline{E_{[t-1]}} * {1-\mu_G}},
where \eqn{\mu_G} corresponds to daily gestating pup mortality.
Gestating Pups transition into nursing pups at the end of \eqn{T_e}. \cr
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
}
