% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_gestReleases_Patch}
\alias{oneDay_gestReleases_Patch}
\title{Determine number of newly conceived mice in a Patch}
\usage{
oneDay_gestReleases_Patch()
}
\description{
Based on this patch's release schedule, \code{\link{generateReleaseVector}},
this function handles daily conception.
}
