% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_nursingDM_deterministic_Patch}
\alias{oneDay_nursingDM_deterministic_Patch}
\title{Deterministic Larva Death and Maturation}
\usage{
oneDay_nursingDM_deterministic_Patch()
}
\description{
Calculate the number of nursing pups surviving from day to day, given by:
\deqn{\overline{L_{[t-1]}} * {1-\mu_N}}
See \code{\link{parameterizeMGDrivE}} for how these parameters are derived.
Maturation has no parameters, so the final day of nursing pups naturally enter the adolescent state.
}
