% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Simulation.R
\name{oneDay_PopDynamics_Patch}
\alias{oneDay_PopDynamics_Patch}
\title{Daily Population Dynamics for a Patch}
\usage{
oneDay_PopDynamics_Patch()
}
\description{
Run population dynamics (including migration) for this patch. \cr
Performed in this order, see the following for each function: \cr
Adult Death: \code{\link{oneDay_adultDeath_deterministic_Patch}} or \code{\link{oneDay_adultDeath_stochastic_Patch}} \cr
Pupa Death/Maturation: \code{\link{oneDay_adoDM_deterministic_Patch}} or \code{\link{oneDay_adoDM_stochastic_Patch}} \cr
Larva Death/Maturation: \code{\link{oneDay_nursingDM_deterministic_Patch}} or \code{\link{oneDay_nursingDM_stochastic_Patch}} \cr
Gestating pup Death/Maturation: \code{\link{oneDay_gestDM_deterministic_Patch}} or \code{\link{oneDay_gestDM_stochastic_Patch}} \cr
Maturation: \code{\link{oneDay_maturation_deterministic_Patch}} or \code{\link{oneDay_maturation_stochastic_Patch}} \cr
Releases: \code{\link{oneDay_releases_Patch}} \cr
Mating: \code{\link{oneDay_mating_deterministic_Patch}} or \code{\link{oneDay_mating_stochastic_Patch}} \cr
Conceive pups: \code{\link{oneDay_conceive_deterministic_Patch}} or \code{\link{oneDay_conceive_stochastic_Patch}} \cr
Release Gestating Pups: \code{\link{oneDay_gestReleases_Patch}} \cr
Migration: \code{\link{oneDay_migrationOut_deterministic_Patch}} or \code{\link{oneDay_migrationOut_stochastic_Patch}} \cr
}
