% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{oneDay_writeOutput_Patch}
\alias{oneDay_writeOutput_Patch}
\title{Write Output from Focal Patch (each day)}
\usage{
oneDay_writeOutput_Patch()
}
\description{
Write Output from Focal Patch (each day)
}
