% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Class.R
\docType{class}
\name{Patch}
\alias{Patch}
\title{Patch Class Definition}
\format{
An \code{\link{R6Class}} generator object
}
\description{
A Patch is a single well-mixed population that is the smallest unit of simulation for MGDrivE.
}
\section{\strong{Constructor}}{

\itemize{
\item patchID: integer ID of this patch
\item genotypesID: character vector of genotypes
\item timeJu: integer vector of length 3 specifying the length of each juvenile stage
\item numPatches: integer, total number of patches in this simulation
\item adultEQ: integer, total adult population in this patch for the duration of the simulation
\item k: double, carrying capacity parameter, see \code{\link{parameterizeMGDrivE}}
\item muAI: additional adult mortality parameter
\item muJI: juvenile infection mortality parameter
\item muN: nursing mortality parameter
\item muG: gestation mortality parameter
\item adultRatioF: named double vector, distribution of adult female genotypes, see \code{\link{parameterizeMGDrivE}}
\item adultRatioM: named double vector, distribution of adult male genotypes, see \code{\link{parameterizeMGDrivE}}
\item gestReleases: gestating pup release schedule for this patch, see \code{\link{basicRepeatedReleases}}
\item maleReleases: male release schedule for this patch, see \code{\link{basicRepeatedReleases}}
\item femaleReleases: female release schedule for this patch, see \code{\link{basicRepeatedReleases}}
\item matedFemaleReleases: mated females release schedule for this patch, see \code{\link{basicRepeatedReleases}}
}
}

\section{\strong{Methods}}{

\itemize{
\item set_NetworkPointer: see \code{\link{set_NetworkPointer_Patch}}
\item get_maleMigration: see \code{\link{get_maleMigration_Patch}}
\item get_femaleMigration: see \code{\link{get_femaleMigration_Patch}}
\item initialPopulation: see \code{\link{set_initialPopulation_Patch}}
\item setPopulation: see \code{\link{set_population_deterministic_Patch}} or \code{\link{set_population_stochastic_Patch}}
\item reset: see \code{\link{reset_Patch}}
\item oneDay_initOutput: see \code{\link{oneDay_initOutput_Patch}}
\item oneDay_writeOutput: see \code{\link{oneDay_writeOutput_Patch}}
\item oneDay_migrationOut: see \code{\link{oneDay_migrationOut_deterministic_Patch}} or \code{\link{oneDay_migrationOut_stochastic_Patch}}
\item oneDay_migrationIn: see \code{\link{oneDay_migrationIn_Patch}}
\item oneDay_PopDynamics: see \code{\link{oneDay_PopDynamics_Patch}}
\item oneDay_adultD: see \code{\link{oneDay_adultDeath_deterministic_Patch}} or \code{\link{oneDay_adultDeath_stochastic_Patch}}
\item oneDay_adoDM: see \code{\link{oneDay_adoDM_deterministic_Patch}} or \code{\link{oneDay_adoDM_stochastic_Patch}}
\item oneDay_nursingDM: see \code{\link{oneDay_nursingDM_deterministic_Patch}} or \code{\link{oneDay_nursingDM_stochastic_Patch}}
\item oneDay_gestDM: see \code{\link{oneDay_gestDM_deterministic_Patch}} or \code{\link{oneDay_gestDM_stochastic_Patch}}
\item oneDay_maturation: see \code{\link{oneDay_maturation_deterministic_Patch}} or \code{\link{oneDay_maturation_stochastic_Patch}}
\item oneDay_releases: see \code{\link{oneDay_releases_Patch}}
\item oneDay_releasePups: see \code{\link{oneDay_gestReleases_Patch}}
\item oneDay_mating: see \code{\link{oneDay_mating_deterministic_Patch}} or \code{\link{oneDay_mating_stochastic_Patch}}
\item oneDay_layPups: see \code{\link{oneDay_oviposit_deterministic_Patch}} or \code{\link{oneDay_oviposit_stochastic_Patch}}
}
}

\section{\strong{Fields}}{

\itemize{
\item patchID: integer ID of this patch
\item popJuvenile: matrix, nGenotype x sum(timeJuvenile), holding all gestating pups, nursing pups, and adolescent pups
\item popMale: vector, nGenotype x 1, holds adult males
\item popFemale: matrix, nGenotype x nGenotype, holds mated adult females
\item popHolder: vector, nGenotype x 1, temporary population storage
\item popPupSex: vector, nGenotype x 1, used in stochastic maturation as another temporary population
\item popUnmated: vector, nGenotype x 1, holds unmated females
\item mMig: matrix, nGenotype x nPatches, holds outbound males for migration, see \code{\link{oneDay_migrationOut_deterministic_Patch}} or \code{\link{oneDay_migrationOut_stochastic_Patch}}
\item fMig: array, nGenotype x nGenotype x nPatches, holds outbound females for migration, see \code{\link{oneDay_migrationOut_deterministic_Patch}} or \code{\link{oneDay_migrationOut_stochastic_Patch}}
\item popJuvenilet0: matrix, nGenotype x sum(timeJuvenile), holding all gestating pups, nursing pups, and maturation for reset, see \code{\link{reset_Patch}}
\item popMalet0: vector, nGenotype x 1, holds adult males for reset see \code{\link{reset_Patch}}
\item popFemalet0: matrix, nGenotype x nGenotype, holds mated adult females for reset see \code{\link{reset_Patch}}
\item gestReleases: list of gestating pup releases for this patch. See \code{\link{oneDay_gestReleases_Patch}}
\item maleReleases: list of adult male releases for this patch. See \code{\link{oneDay_releases_Patch}}
\item femaleReleases: list of adult female releases for this patch. See \code{\link{oneDay_releases_Patch}}
\item matedFemaleReleases: list of mated adult female releases for this patch. See \code{\link{oneDay_releases_Patch}}
\item NetworkPointer: a reference to enclosing \code{\link{Network}}
}
}

\keyword{R6}
\keyword{class}
