% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{set_initialPopulation_Patch}
\alias{set_initialPopulation_Patch}
\title{Set Initial Population}
\usage{
set_initialPopulation_Patch(
  k = k,
  adultRatioF = adultRatioF,
  adultRatioM = adultRatioM,
  muAI = muAI,
  muJI = muJI,
  muN = muN,
  muG = muG,
  timeAd = timeAd
)
}
\arguments{
\item{k}{Carrying capacity constant}

\item{adultRatioF}{Genotype specific ratio for adult females}

\item{adultRatioM}{Genotype specific ratio for adult males}

\item{muAI}{daily survival of adult stage}

\item{muJI}{juvenile infection mortality parameter}

\item{muN}{nursing mortality parameter}

\item{muG}{gestation mortality parameter}

\item{timeAd}{approximate length of adult life stage}
}
\description{
This hidden function distributes the population at time 0 in the steady-state
conformation. This involves splitting adults into male and female.
}
