% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patch-Methods.R
\name{set_population_stochastic_Patch}
\alias{set_population_stochastic_Patch}
\title{Set Initial Population (Stochastic)}
\usage{
set_population_stochastic_Patch(
  k = k,
  adultRatioF = adultRatioF,
  adultRatioM = adultRatioM,
  use_eta_at_t0 = FALSE,
  ...
)
}
\description{
Set Initial Population (Stochastic)
}
